move generateRangePicker out
