+++ components/date-picker/generatePicker/index.tsx
@@ -1,6 +1,4 @@
 import * as React from 'react';
-import classNames from 'classnames';
-import { RangePicker as RCRangePicker } from 'rc-picker';
 import { GenerateConfig } from 'rc-picker/lib/generate/index';
 import {
   PickerBaseProps as RCPickerBaseProps,
@@ -14,18 +12,12 @@ import {
   RangePickerTimeProps as RCRangePickerTimeProps,
 } from 'rc-picker/lib/RangePicker';
 import { PickerMode, Locale as RcPickerLocale } from 'rc-picker/lib/interface';
-import CalendarOutlined from '@ant-design/icons/CalendarOutlined';
-import ClockCircleOutlined from '@ant-design/icons/ClockCircleOutlined';
-import CloseCircleFilled from '@ant-design/icons/CloseCircleFilled';
-import { ConfigContext, ConfigConsumerProps } from '../../config-provider';
-import LocaleReceiver from '../../locale-provider/LocaleReceiver';
-import enUS from '../locale/en_US';
-import { getRangePlaceholder } from '../util';
+import { SizeType } from '../../config-provider/SizeContext';
 import PickerButton from '../PickerButton';
 import PickerTag from '../PickerTag';
-import SizeContext, { SizeType } from '../../config-provider/SizeContext';
 import { TimePickerLocale } from '../../time-picker';
 import generateSinglePicker from './generateSinglePicker';
+import generateRangePicker from './generateRangePicker';
 
 export const Components = { button: PickerButton, rangeItem: PickerTag };
 
@@ -136,102 +128,7 @@ function generatePicker<DateType>(generateConfig: GenerateConfig<DateType>) {
   );
 
   // ======================== Range Picker ========================
-  class RangePicker extends React.Component<RangePickerProps<DateType>> {
-    static contextType = ConfigContext;
-
-    context: ConfigConsumerProps;
-
-    pickerRef = React.createRef<RCRangePicker<DateType>>();
-
-    focus = () => {
-      if (this.pickerRef.current) {
-        this.pickerRef.current.focus();
-      }
-    };
-
-    blur = () => {
-      if (this.pickerRef.current) {
-        this.pickerRef.current.blur();
-      }
-    };
-
-    getDefaultLocale = () => {
-      const { locale } = this.props;
-      const result = {
-        ...enUS,
-        ...locale,
-      };
-      result.lang = {
-        ...result.lang,
-        ...((locale || {}) as PickerLocale).lang,
-      };
-      return result;
-    };
-
-    renderPicker = (locale: PickerLocale) => {
-      const { getPrefixCls, direction } = this.context;
-      const {
-        prefixCls: customizePrefixCls,
-        className,
-        size: customizeSize,
-        bordered = true,
-        ...restProps
-      } = this.props;
-      const { format, showTime, picker } = this.props as any;
-      const prefixCls = getPrefixCls('picker', customizePrefixCls);
-
-      let additionalOverrideProps: any = {};
-
-      additionalOverrideProps = {
-        ...additionalOverrideProps,
-        ...(showTime ? getTimeProps({ format, picker, ...showTime }) : {}),
-        ...(picker === 'time' ? getTimeProps({ format, ...this.props, picker }) : {}),
-      };
-
-      return (
-        <SizeContext.Consumer>
-          {size => {
-            const mergedSize = customizeSize || size;
-
-            return (
-              <RCRangePicker<DateType>
-                separator={<span aria-label="to" className={`${prefixCls}-separator`} />}
-                ref={this.pickerRef}
-                placeholder={getRangePlaceholder(picker, locale)}
-                suffixIcon={picker === 'time' ? <ClockCircleOutlined /> : <CalendarOutlined />}
-                clearIcon={<CloseCircleFilled />}
-                allowClear
-                transitionName="slide-up"
-                {...restProps}
-                className={classNames(className, {
-                  [`${prefixCls}-${mergedSize}`]: mergedSize,
-                  [`${prefixCls}-borderless`]: !bordered,
-                })}
-                {...additionalOverrideProps}
-                locale={locale!.lang}
-                prefixCls={prefixCls}
-                generateConfig={generateConfig}
-                prevIcon={<span className={`${prefixCls}-prev-icon`} />}
-                nextIcon={<span className={`${prefixCls}-next-icon`} />}
-                superPrevIcon={<span className={`${prefixCls}-super-prev-icon`} />}
-                superNextIcon={<span className={`${prefixCls}-super-next-icon`} />}
-                components={Components}
-                direction={direction}
-              />
-            );
-          }}
-        </SizeContext.Consumer>
-      );
-    };
-
-    render() {
-      return (
-        <LocaleReceiver componentName="DatePicker" defaultLocale={this.getDefaultLocale}>
-          {this.renderPicker}
-        </LocaleReceiver>
-      );
-    }
-  }
+  const RangePicker = generateRangePicker(generateConfig);
 
   // =========================== Export ===========================
   type MergedDatePicker = typeof DatePicker & {
